requirements.txt
openai>=1.0.0
pyyaml>=6.0
defusedxml>=0.7.0
ruff>=0.1.0
